+++ spec/lib/votd/netbible_spec.rb
@@ -55,6 +55,12 @@ describe "Votd::NETBible" do
     end
   end
 
+  describe ".link" do
+    it 'returns the link' do
+      expect(votd.link).to eq 'https://netbible.org/bible/Ephesians+2:8'
+    end
+  end
+
   describe ".to_html" do
     it "returns a HTML version" do
       expect(votd.to_html).to eq read_fixture("netbible/netbible.html")